Split the make-token command out of brkt_jwt
Split the code that handles the make-token subcommand out of the
brkt_jwt module.  This divides responsibility more clearly, and avoids
circular dependencies in JWT library callsites (e.g.
callsite->brkt_jwt->brkt_cli->brkt_jwt).

Also move the timestamp parsing functions into the util module, since
they're called from multiple places.
