+++ brkt_cli/brkt_jwt/test_jwt.py
@@ -11,42 +11,23 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and
 # limitations under the License.
-import time
 import unittest
 import uuid
 from datetime import datetime, timedelta
 
 import iso8601
+import jwt as pyjwt
 
 import brkt_cli.crypto
-import brkt_cli.util
+from brkt_cli import util
 from brkt_cli import brkt_jwt
 from brkt_cli.crypto import test_crypto
 from brkt_cli.validation import ValidationError
-import jwt as pyjwt
 
 if brkt_cli.crypto.cryptography_library_available:
     _crypto = brkt_cli.crypto.from_private_key_pem(
         test_crypto.TEST_PRIVATE_KEY_PEM
     )
-
-
-class TestTimestamp(unittest.TestCase):
-
-    def test_datetime_to_timestamp(self):
-        now = time.time()
-        dt = datetime.fromtimestamp(now, tz=iso8601.UTC)
-        self.assertEqual(now, brkt_jwt._datetime_to_timestamp(dt))
-
-    def test_parse_timestamp(self):
-        ts = int(time.time()) + 10     # JWT timestamps must be in the future
-        dt = datetime.fromtimestamp(ts, tz=iso8601.UTC)
-
-        self.assertEqual(dt, brkt_jwt.parse_timestamp(str(ts)))
-        self.assertEqual(dt, brkt_jwt.parse_timestamp(dt.isoformat()))
-
-        with self.assertRaises(ValidationError):
-            brkt_jwt.parse_timestamp('abc')
 
 
 class TestGenerateJWT(unittest.TestCase):
@@ -88,13 +69,13 @@ class TestGenerateJWT(unittest.TestCase):
         self.assertEqual(1, payload['one'])
         self.assertEqual(2, payload['two'])
 
-        iat = brkt_jwt._timestamp_to_datetime(payload['iat'])
+        iat = util.timestamp_to_datetime(payload['iat'])
         self.assertTrue(now <= iat <= after)
 
-        nbf_result = brkt_jwt._timestamp_to_datetime(payload['nbf'])
+        nbf_result = util.timestamp_to_datetime(payload['nbf'])
         self.assertEqual(nbf, nbf_result)
 
-        exp_result = brkt_jwt._timestamp_to_datetime(payload['exp'])
+        exp_result = util.timestamp_to_datetime(payload['exp'])
         self.assertEqual(exp, exp_result)
 
     def test_malformed(self):
@@ -139,10 +120,10 @@ class TestNameValueToDict(unittest.TestCase):
     def test_name_value_to_dict(self):
         self.assertEqual(
             {'a': 'b', 'c': 'd'},
-            brkt_jwt._name_value_list_to_dict(['a=b', 'c=d'])
+            brkt_jwt.name_value_list_to_dict(['a=b', 'c=d'])
         )
         with self.assertRaises(ValidationError):
-            brkt_jwt._name_value_list_to_dict(['a=b', 'a=c'])
+            brkt_jwt.name_value_list_to_dict(['a=b', 'a=c'])
 
     def test_brkt_tags_from_name_value_list(self):
         self.assertEqual(
